# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Стильный призыв API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /content:
    get:
      operationId: listContent
      tags: [content]
      summary: List all site content items
      responses:
        "200":
          description: Content items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListContentPayload"
    put:
      operationId: upsertContent
      tags: [content]
      summary: Upsert content items (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertContentInput"
      responses:
        "200":
          description: Updated content items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpsertContentPayload"
  /content/{key}:
    delete:
      operationId: deleteContent
      tags: [content]
      summary: Delete a content item (owner only)
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorPayload"

  # ── Posts ──────────────────────────────────────────────────────────────
...
  /hadiths:
    get:
      operationId: listHadiths
      tags: [hadiths]
      summary: List published hadiths
      parameters:
        - name: grade
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/HadithGrade"
        - name: topic
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Hadiths
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListHadithsPayload"
    post:
      operationId: createHadith
      tags: [hadiths]
      summary: Create a hadith (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HadithInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateHadithPayload"
...
  /hadiths/{id}:
    get:
      operationId: getHadith
      tags: [hadiths]
      summary: Get a single published hadith
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Hadith
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetHadithPayload"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorPayload"
    patch:
      operationId: updateHadith
      tags: [hadiths]
      summary: Update a hadith (owner, or editor for own hadith)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HadithUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateHadithPayload"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorPayload"
    delete:
      operationId: deleteHadith
      tags: [hadiths]
      summary: Delete a hadith (owner, or editor for own hadith)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorPayload"

  # ── Admins ─────────────────────────────────────────────────────────────
...
  /hadiths/{id}/interactions:
    get:
      operationId: getHadithInteractions
      tags: [hadith-interactions]
      summary: Get comments and reaction counts for a hadith
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Interactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HadithInteractions"
  /hadiths/{id}/comments:
    post:
      operationId: createHadithComment
      tags: [hadith-interactions]
      summary: Add a comment to a hadith (signed-in users only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HadithComment"
  /hadiths/{id}/comments/{commentId}:
    delete:
      operationId: deleteHadithComment
      tags: [hadith-interactions]
      summary: Delete a hadith comment (own comment or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: commentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /hadiths/{id}/react:
    post:
      operationId: reactToHadith
      tags: [hadith-interactions]
      summary: Like or dislike a hadith (toggle, signed-in users only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionInput"
      responses:
        "200":
          description: Current reaction state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactToHadithPayload"

  # ── User profiles ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorPayload:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ContentItem:
      type: object
      required: [id, key, group, label, type, value, order, updatedAt]
      properties:
        id: { type: integer }
        key: { type: string }
        group: { type: string }
        label: { type: string }
        type:
          type: string
          enum: [text, textarea, url, color, image]
        value: { type: string }
        order: { type: integer }
        updatedAt: { type: string }
    ListContentPayload:
      type: array
      items:
        $ref: "#/components/schemas/ContentItem"
    ContentItemInput:
      type: object
      required: [key, value]
      properties:
        key: { type: string }
        group: { type: string }
        label: { type: string }
        type:
          type: string
          enum: [text, textarea, url, color, image]
        value: { type: string }
        order: { type: integer }
    UpsertContentInput:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ContentItemInput"
    UpsertContentPayload:
      type: array
      items:
        $ref: "#/components/schemas/ContentItem"

    # ── Posts ──────────────────────────────────────────────────────────
...
    CommentInput:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1 }
    ReactionInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [like, dislike]
...
    HadithGrade:
      type: string
      enum: [sahih, hasan, daif, mawdu]
    Hadith:
      type: object
      required: [id, text, source, narrator, grade, topic, published, authorName, createdAt, updatedAt]
      properties:
        id: { type: integer }
        text: { type: string }
        source: { type: string }
        narrator: { type: string }
        grade:
          $ref: "#/components/schemas/HadithGrade"
        topic: { type: string }
        published: { type: boolean }
        authorId: { type: ["integer", "null"] }
        authorName: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
    ListHadithsPayload:
      type: array
      items:
        $ref: "#/components/schemas/Hadith"
...
    HadithInput:
      type: object
      required: [text, grade]
      properties:
        text: { type: string, minLength: 1 }
        source: { type: string }
        narrator: { type: string }
        grade:
          $ref: "#/components/schemas/HadithGrade"
        topic: { type: string }
        published: { type: boolean }
    HadithUpdate:
      type: object
      properties:
        text: { type: string, minLength: 1 }
        source: { type: string }
        narrator: { type: string }
        grade:
          $ref: "#/components/schemas/HadithGrade"
        topic: { type: string }
        published: { type: boolean }
    GetHadithPayload:
      $ref: "#/components/schemas/Hadith"
    CreateHadithPayload:
      $ref: "#/components/schemas/Hadith"
    UpdateHadithPayload:
      $ref: "#/components/schemas/Hadith"

    # ── Hadith interactions ────────────────────────────────────────────
    HadithComment:
      type: object
      required: [id, hadithId, authorName, isAdmin, content, createdAt]
      properties:
        id: { type: integer }
        hadithId: { type: integer }
        authorClerkId: { type: ["string", "null"] }
        authorName: { type: string }
        authorAvatarUrl: { type: ["string", "null"] }
        isAdmin: { type: boolean }
        content: { type: string }
        createdAt: { type: string }
    HadithInteractions:
      type: object
      required: [comments, likes, dislikes, myReaction]
      properties:
        comments:
          type: array
          items:
            $ref: "#/components/schemas/HadithComment"
        likes: { type: integer }
        dislikes: { type: integer }
        myReaction:
          type: ["string", "null"]
          enum: [like, dislike, null]
    ReactToHadithPayload:
      type: object
      required: [myReaction]
      properties:
        myReaction:
          type: ["string", "null"]
          enum: [like, dislike, null]

    # ── User profiles ──────────────────────────────────────────────────
...